Move some files around
This will be useful once we start splitting out "systems" and "backends"
a bit more, e.g. when polling support is added.
