+++ system_darwin_test.go
@@ -1,3 +1,6 @@
+//go:build darwin
+// +build darwin
+
 package fsnotify
 
 import (
